feat: reselect settings selectorMethodPrefix and stateParamName
`reselect/selectorMethodPrefix` setting
Specify a non-'get' selector method name prefix, so all rules recognize alternative selector method naming conventions

`reselect/stateParamName` setting
Specify a non-'state' param name for `call` and `first-param-name` rules.
Because `call`'s condition pattern includes an implicit check against a `state` param, this setting provides a single spot in which the name of this param is declared, and which applies to both the `call` and `first-param-name` rules.